+++ subsequences/rabi_excitation.py
@@ -1,4 +1,5 @@
 from artiq.experiment import *
+import numpy as np
 
 class RabiExcitation:
     freq_729="Excitation_729.rabi_excitation_frequency"
@@ -10,7 +11,7 @@ class RabiExcitation:
     line_selection="Excitation_729.line_selection"
     sp_amp_729="Excitation_729.single_pass_amplitude"
     sp_att_729="Excitation_729.single_pass_att"
-    phase_ref_time=0
+    phase_ref_time=np.int64(0)
 
     def subsequence(self):
         r = RabiExcitation
